cahnged type from double to decimal

+++ ExpenseProcessingSystem/ExpenseProcessingSystem/Controllers/ModalController.cs
@@ -1017,24 +1017,24 @@ namespace ExpenseProcessingSystem.Controllers
 
             List<SelectListItem> currList = _service.getCurrencyIDSelectList(int.Parse(Curr1AbbrID));
             Curr2AbbrID = Curr2AbbrID == "0" ? currList.Select(x => x.Value).FirstOrDefault() : Curr2AbbrID;
-            List<InterEntityParticular> parList1 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular1(account, Curr1AbbrName, float.Parse(Curr1Amt), float.Parse(Curr2Amt), InterRate, int.Parse(accID), int.Parse(Curr1AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry1']"));
-            List<InterEntityParticular> parList2 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular2(Curr1AbbrName, Curr2AbbrName, float.Parse(Curr2Amt), InterRate, int.Parse(Curr1AbbrID), int.Parse(Curr2AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry2']"));
-            List<InterEntityParticular> parList3 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular3(Curr2AbbrName, float.Parse(Curr2Amt), int.Parse(Curr2AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry3']"));
+            List<InterEntityParticular> parList1 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular1(account, Curr1AbbrName, decimal.Parse(Curr1Amt), decimal.Parse(Curr2Amt), InterRate, int.Parse(accID), int.Parse(Curr1AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry1']"));
+            List<InterEntityParticular> parList2 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular2(Curr1AbbrName, Curr2AbbrName, decimal.Parse(Curr2Amt), InterRate, int.Parse(Curr1AbbrID), int.Parse(Curr2AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry2']"));
+            List<InterEntityParticular> parList3 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular3(Curr2AbbrName, decimal.Parse(Curr2Amt), int.Parse(Curr2AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry3']"));
             
             DDVInterEntityViewModel model = new DDVInterEntityViewModel
             {
                 Inter_ID = int.Parse(id.Substring(5)),
                 Inter_Currency1_ID = int.Parse(Curr1AbbrID),
                 Inter_Currency1_ABBR = Curr1AbbrName,
-                Inter_Currency1_Amount = float.Parse(Curr1Amt),
+                Inter_Currency1_Amount = decimal.Parse(Curr1Amt),
                 Inter_Currency2_ID = Curr2AbbrID == "0" ? int.Parse(currList.Select(x => x.Value).FirstOrDefault()) : int.Parse(Curr2AbbrID),
                 Inter_Currency2_ABBR = Curr2AbbrID == "0" ? currList.Select(x => x.Text).FirstOrDefault() : Curr2AbbrName,
-                Inter_Currency2_Amount = float.Parse(Curr2Amt),
+                Inter_Currency2_Amount = decimal.Parse(Curr2Amt),
                 Inter_Rate = InterRate,
                 Inter_Check1 = Chk1 == "True" ? true : Chk1 == "true" ? true : false,
                 Inter_Check2 = Chk2 == "True" ? true : Chk2 == "true" ? true : false,
-                Inter_Convert1_Amount = float.Parse(Conv1Amt),
-                Inter_Convert2_Amount = float.Parse(Conv2Amt),
+                Inter_Convert1_Amount = decimal.Parse(Conv1Amt),
+                Inter_Convert2_Amount = decimal.Parse(Conv2Amt),
                 interPartList = new List<ExpenseEntryInterEntityParticularViewModel>(),
                 CurrencyList = currList
             };
@@ -1055,9 +1055,9 @@ namespace ExpenseProcessingSystem.Controllers
             float InterRate = interRate != null ? float.Parse(interRate) : 1;
             string remarks = remarksTitle + " " + ConstantData.HomeReportConstantValue.GetMonthList().Where(c => c.MonthID == DateTime.Now.Month).Single().MonthName + " " + DateTime.Now.Year;
 
-            List<InterEntityParticular> parList1 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular1(account, Curr1AbbrName, float.Parse(Curr1Amt), float.Parse(Curr2Amt), InterRate, int.Parse(accID), int.Parse(Curr1AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry1']"));
-            List<InterEntityParticular> parList2 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular2(Curr1AbbrName, Curr2AbbrName, float.Parse(Curr2Amt), InterRate, int.Parse(Curr1AbbrID), int.Parse(Curr2AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry2']"));
-            List<InterEntityParticular> parList3 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular3(Curr2AbbrName, float.Parse(Curr2Amt), int.Parse(Curr2AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry3']"));
+            List<InterEntityParticular> parList1 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular1(account, Curr1AbbrName, decimal.Parse(Curr1Amt), decimal.Parse(Curr2Amt), InterRate, int.Parse(accID), int.Parse(Curr1AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry1']"));
+            List<InterEntityParticular> parList2 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular2(Curr1AbbrName, Curr2AbbrName, decimal.Parse(Curr2Amt), InterRate, int.Parse(Curr1AbbrID), int.Parse(Curr2AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry2']"));
+            List<InterEntityParticular> parList3 = CONSTANT_DDV_INTER_PARTICULARS.PopulateParticular3(Curr2AbbrName, decimal.Parse(Curr2Amt), int.Parse(Curr2AbbrID), _service.getInterEntityAccs("/INTERENTITYACCOUNTS/ACCOUNT[@class='entry3']"));
             List<SelectListItem> currList = _service.getCurrencySelectList();
 
             DDVInterEntityViewModel model = new DDVInterEntityViewModel
@@ -1065,15 +1065,15 @@ namespace ExpenseProcessingSystem.Controllers
                 Inter_ID = int.Parse(id.Substring(5)),
                 Inter_Currency1_ID = int.Parse(Curr1AbbrID),
                 Inter_Currency1_ABBR = Curr1AbbrName,
-                Inter_Currency1_Amount = float.Parse(Curr1Amt),
+                Inter_Currency1_Amount = decimal.Parse(Curr1Amt),
                 Inter_Currency2_ID = Curr2AbbrID == "NAN" ? 2 : int.Parse(Curr2AbbrID),
                 Inter_Currency2_ABBR = Curr2AbbrID == "NAN" ? _context.DMCurrency.Where(x => x.Curr_MasterID == 2 && x.Curr_isActive == true && x.Curr_isDeleted == false).Select(x => x.Curr_CCY_ABBR).FirstOrDefault() : Curr2AbbrName,
-                Inter_Currency2_Amount = float.Parse(Curr2Amt),
+                Inter_Currency2_Amount = decimal.Parse(Curr2Amt),
                 Inter_Rate = InterRate,
                 Inter_Check1 = Chk1 == "True" ? true : Chk1 == "true" ? true : false,
                 Inter_Check2 = Chk2 == "True" ? true : Chk2 == "true" ? true : false,
-                Inter_Convert1_Amount = float.Parse(Conv1Amt),
-                Inter_Convert2_Amount = float.Parse(Conv2Amt),
+                Inter_Convert1_Amount = decimal.Parse(Conv1Amt),
+                Inter_Convert2_Amount = decimal.Parse(Conv2Amt),
                 interPartList = new List<ExpenseEntryInterEntityParticularViewModel>(),
                 CurrencyList = currList
             };
@@ -1089,7 +1089,7 @@ namespace ExpenseProcessingSystem.Controllers
         }
         //_________________________//[Petty Cash/Cash Advance(Suspense sundry) Expense]//_____________________________
         //Expense Cash Breakdown
-        public IActionResult EntryExpenseCashBreakdown(string id, string vendor, string account, double amount, string screencode)
+        public IActionResult EntryExpenseCashBreakdown(string id, string vendor, string account, decimal amount, string screencode)
         {
             PCVCashBreakdownViewModel model = new PCVCashBreakdownViewModel();
 
@@ -1180,13 +1180,13 @@ namespace ExpenseProcessingSystem.Controllers
                 model.CBD_one = pc.PCB_One;
                 model.CBD_oneAmount = pc.PCB_One * 1;
                 model.CBD_c25 = pc.PCB_TwentyFiveCents;
-                model.CBD_c25Amount = pc.PCB_TwentyFiveCents * .25;
+                model.CBD_c25Amount = (decimal)(pc.PCB_TwentyFiveCents * .25);
                 model.CBD_c10 = pc.PCB_TenCents;
-                model.CBD_c10Amount = pc.PCB_TenCents * .10;
+                model.CBD_c10Amount = (decimal)(pc.PCB_TenCents * .10);
                 model.CBD_c5 = pc.PCB_FiveCents;
-                model.CBD_c5Amount = pc.PCB_FiveCents * .05;
+                model.CBD_c5Amount = (decimal)(pc.PCB_FiveCents * .05);
                 model.CBD_c1 = pc.PCB_OneCents;
-                model.CBD_c1Amount = pc.PCB_OneCents * .01;
+                model.CBD_c1Amount = (decimal)(pc.PCB_OneCents * .01);
             }
             else
             {
